Prefer app directories on removable storage
Since the application doesn't require any special privileges to access
data in the app private data directory, look for the Endless Key content
there before falling back to the root of the removable storage volume.

This depends on the Endless Key USB being built with the `KOLIBRI_DATA`
directory in `Android/data/org.endlessos.Key/files` rather than in the
root of the volume.

https://phabricator.endlessm.com/T33663