Fully customize logging
Provide a full logging dictionary configuration and wrapper function
replacing `kolibri.utils.logger.get_default_config`. This has 3
advantages:

1. Allows adding `AndroidLogHandler` on the root logger so that python
   logging records are nicely output to the Android log.

2. Since we define the `RotatingFileHandler` immediately, all logs will
   be output to the `logs/kolibri.txt` file. That may help debugging
   issues later since the Android log is only in memory.

3. Ensures Kolibri uses the same logging configuration and doesn't
   replace our handlers or duplicate logs output to stdout or stderr.

Unfortunately, the only way to currently do that in Kolibri is via
monkeypatching. Specifying a custom Django `LOG` setting is possible,
but that comes in quite a bit later when Django is initialized. This
ensures we have a single logging configuration as soon as the python
code starts running.