Update logcat usage
With the customized logging, most of the interesting logs won't be under
the `python` tag anymore. The Java classes also have different tags that
would be missed by common filtering. While here, delete `logcat.sh`,
which likely wasn't used by anyone since it was filtering on the wrong
tags. It's really not needed with proper tagging.