Log exception through logging system
The default `sys.excepthook` (also available as `sys.__excepthook__` for
future reference) formats an exception and prints it to stderr. This
eventually shows up in the Android logging system through
pythonforandroid's stderr to Log converter, but that suffers all the
same problems that `AndroidLogHandler` addresses. Furthermore, since the
exception is sent directly to stderr, it doesn't get recorded in the
`kolibri.txt` file through its logging handler. Recording it with
`logging.critical` means it goes to both places with proper formatting
and priority.