Allow passing a local build of kolibri-explore-plugin
For debugging purposes. If the EXPLOREPLUGIN_WHEEL_PATH environment
variable is passed, it is installed in src/ which overrides the one
installed by python-for-android through the .p4a --requirements option.
Also the welcomeScreen asset is copied from it.

If not passed, the welcomeScreen asset is still copied from PyPi by
installing the kolibri-explore-plugin package in a temporary directory.