jenkins: Add default value to RELEASE_NOTES of kolibri-installer-android-upload
We use kolibri-installer-android-upload job only to upload to internal
testers. The real release notes are written in Google Play Console and
only when promoting the testing build. For internal testers the notes
aren't actually needed.

Add an empty array as the default value to the parameter RELEASE_NOTES
of kolibri-installer-android-upload job.

https://phabricator.endlessm.com/T34808