jenkins: Don't set parameters in pipelines
We manage job parameters with jenkins-job-builder, so every time it runs
it will reset whatever parameters are defined in the pipeline. While
it's nice to declare the parameters in the pipeline and therefore
guarantee that they'll be defined, it means that the definition has to
exactly match what JJB defines or they'll contstantly bounce back and
forth. Instead, just add a comment about what parameters are required.

https://phabricator.endlessm.com/T34805