jenkins: Select current build job when triggering upload
If the upload job is being triggered by the build job, it will not be
the last successful build yet and `copyArtifacts` will use the AAB from
the previous build. Instead, use the `upstream` selector[1], which uses
the build number of the upstream (triggering) job. Since the upload job
can still be run manually, fallback to the last successful build job.

1. https://plugins.jenkins.io/copyartifact/#plugin-content-pipeline-syntax

https://phabricator.endlessm.com/T34805