Handle outdated webview versions
If the system webview package is too old, we're likely to crash later
using unsupported JS and web platform APIs. If the detected webview
package is too old, start a dialog the instructs the user and opens the
Play Store to the package's page.

At the moment this only handles `com.google.android.webview`, which is
the standard Android webview. That's likely all we'll see on
Chromebooks.

Fixes: #160