Don't cover status bar on Android
By default python-for-android appends `.Fullscreen` to the theme. On
ChromeOS this has no effect, but on Android it makes the app cover the
status bar. I don't believe Endless Key is so important that it needs
that and it introduces a difference with ChromeOS where the app always
runs windowed and the status bar color is used for the window
decoration. Add the p4a `--window` option so that it doesn't append
`.Fullscreen` to the theme.