Run task workers in separate process bus
Running Kolibri tasks cannot be cancelled[1] and shutting down the
process bus running them will block until the running tasks complete.
This means the Kolibri server can't be stopped quickly or reliably
during the activity's `onActivityStopped` hook.

To workaround that, move the task workers spawned from the
`ServicesPlugin` into a separate process bus. This bus will only be
stopped when the activity is destroyed (or killed). Besides making the
server lifecycle much more reliable, it allows tasks to continue running
and starting in the background as long as the activity process is alive.

Ideally the server and workers would be separate services outside of the
main webview activity, but python-for-android makes that really hard to
accomplish.

1. https://github.com/learningequality/kolibri/issues/10249

Helps: https://github.com/endlessm/kolibri-explore-plugin/issues/740