build: Bump to Android API 33
This is needed because Google now require that all apps target an API
level within 1 year of the latest Android release.

The API changes from 31 → 32 and 32 → 33 are minimal and should not
affect our code:
 * https://developer.android.com/sdk/api_diff/32/changes
 * https://developer.android.com/sdk/api_diff/33/changes

It requires a newer version of the NDK and build tools. NDK 25 is the
lowest version of the NDK which supports API 33, so let’s go with that.

There are no `default` system image builds for API 33, so switch to the
most appropriate next choice, which looks like `google_apis_playstore`.

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>

Fixes: https://github.com/endlessm/kolibri-explore-plugin/issues/759