Revert "Run task workers in separate process bus"
This reverts commit 4d22c7c647f962db91e11068fed8d93fa27cc1f4. With
kolibri 0.16.0-beta5, tasks can be stopped and there's less reason to
try to split those up. Unlike `onDestroy`, `onStop` is always delivered,
so this gives Kolibri a chance to stop.

In theory it would be better to keep the process buses separate so that
the server and worker run independently. However, that would only really
be a benefit if they could be run in separate bound services. That's not
supported in python-for-android, so for now it's better to run it all
from the single python thread in the activity.