Initial gradle build infrastructure
This begins the process of building the app natively using Gradle
instead of having python-for-android manage the build. It's setup as a
multi-project with the app as its own project since this is the
customary way Android builds are setup.

The app will build successfully, but it doesn't actually do anything
since there aren't any activities or services defined.