Generate version numbers in gradle
This is a little more complicated since the versions depend on the
downloaded packages and chaquopy is handling that. A task is inserted
after chaquopy extracts the packages and hooked into AGP's `onVariants`.

The version script is changed to output the JSON file used in
Jenkinsfile-upload. Since Gradle uses input and output files to detect
when a task has changed, this also has the benefit of ensuring the
versions are updated whenever packages are extracted.