Add pre-commit configuration for Java and Kotlin
These will be much more heavily used with chaquopy. I chose the AOSP
Java style instead of the regular Google format since we're writing
Android code. While here:

* Update the existing Java code to match the accepted style.
* Add an EditorConfig file is added to hopefully get your editor to more
  closely match the expected style.
* Add pre-commit to the python requirements and run it during Jenkins
  builds.

The github workflow is dropped since it duplicates the work in Jenkins.
Maybe later we'll move everything to github actions, but for now the
real work is in Jenkins.