Extract apps bundle and collections into python source directory
The apps bundle and collections files need to be accessible on disk from
the explore plugin code. The most appropriate way to do that would be to
store the zip files as assets and expand them at runtime. However,
chaquopy already does that for all data files, so just put them in the
python package.

Both to ensure they're extracted when importing `kolibri_android` and
because it's more appropriate, put the data files (including the
automatic provision file) in the package directory rather than the root
of the source set.