gradle: Add Firebase Crashlytics
Like analytics, currently this is simply enabling or disabling analytics
based on whether it's a release build or not. Later the runtime
enablement will be ported from the python code. This is also missing the
pieces for the native symbol upload (which we never had before, either).

As a bonus, since chaquopy already converts python exceptions its
`PyException` java class far better than we ever did, we can drop all
of our global exception proxying.