Revert "Disable analytics for Endless testing installs"
This reverts commit ee22c8b591d80945208f10db8f126caab9e364d4. It turns
out that when the user is logged in with a Google Workspace account, the
install referrer is always set to
`utm_source=(not%20set)&utm_medium=(not%20set)`. That's not helpful and
defeats the purpose of using the install referrer to guide analytics
since most Endless employees will be logged in from a Google Workspace
account.

Helps: #158