Move kolibri_android python code to chaquopy source set
This is not wired up yet and will require several changes, but this is
where it will live to be accessed from java. Note that the script entry
points are not moved over as chaquopy allows calling python attributes
rather than running a single python script until the component ends.