Clean up share file interface
The intention from the Kolibri side is that this is for sharing files,
not messages. The point of the message is to accompany provide context
in the OS sharing implementation. The only use for the message in this
case is to use it as a title on an activity chooser. See the
`ACTION_SEND`[1] documentation for details.

As it is, none of this is used in Kolibri right now, so this is just
academic.

1. https://developer.android.com/reference/android/content/Intent#ACTION_SEND