Add file provider class
The previous implementation purely in python would not have worked
because the generic `FileProvider` class does not serve the
`org.endlessos.Key.fileprovider` authority.

Instead, a subclass is created, a `provider` is added to the manifest,
and a convenience static method is added to construct the appropriate
intent from python.