Add Kolibri service
Run Kolibri as a service which is bound from the main webview activity.
The service is bound and unbound in the activity's start and stop
states. Since the service runs in a separate process, this gives Kolibri
a chance to stop gracefully when the activity is killed. In order for
the activity to find the current URL and app key, a small `Message`
based IPC is used.

Chaquopy expects to return from Python calls, so rather than blocking on
the server, the process bus is started and stopped as the service is
created and destroyed. This requires a bit of refactoring in the python
code.

Similarly, since we have full access to the activity and service
lifecycle hooks, a lot of the python code can go away as it simply needs
to start the Kolibri process bus.