Stop disabling Gradle daemon in container
We now run Gradle multiple times from the container, so letting it start
a daemon does improve the overall speed. Furthermore, even if you say
not to use a daemon, Gradle 8 starts a single use daemon since it's part
of the implementation. Like the comment previously said, any daemon will
be killed as soon as the container is stopped.