Add version code back to version name
This was lost in 0b0edae2729a21372fc4be5c3cb2e25a979ca62f when moving
the version handling script to use as a Gradle task. Initially it didn't
have access to the version code, so I just dropped that from the version
name. However, later I allowed passing the version code as an option to
simplify generation of the version JSON file. Since the version code may
be available, restore the `-$version_code` suffix in the version name if
possible.