gradle: Sync zip asset contents
`Copy` tasks (and the `Project.copy` method) add files to the
destination directory, but that means old files in the tree are still
included. That caused the last build to exceed the 150 MB limit despite
a slimmer `apps-bundle.zip` in the last explore plugin release.

Instead, use `Sync` tasks (and the `Project.sync` method) to ensure that
only the files extracted from the zip assets are included.